Update tokenizer.py to decode list of tensors as well as single tensor
This is an enhancement to fix functionality of Lollms-webui, specifically the lord of internet and wikipedia personalities, as well as anything else that uses workflows or returns list of tensors to be decoded. 

This change identifies the type, and decodes based on the type. If a list is detected it calls itself recursively and decodes each item in the list until completion. 

https://github.com/ParisNeo/lollms-webui/issues/480#issuecomment-1903125058