dynamic_async: Properly close the iterator loop on exit (#538)
When the close method is called, the generator's iterator loop never
actually exited. This is because the condition is not notified meaning
the task is still running even though it's signalled to cancel.

Therefore, add an extra pass if the task is cancelled and unlock the
loop by forcing a notify on close. From there, normal cancellation
handling will work.

There might be a better way to do this, but this way minimizes the amount
of added code and makes the most(?) sense.

Signed-off-by: kingbri <bdashore3@proton.me>